Extract Breakpoint component
This is a simple extraction. There are a couple reasons for that:

1. I'd like sufficiently independent components to be in their own file.

2. The editor component will be one of our most complicated components,
so its good to think about ways that we can simplify the file so that it
is easier to understand. Maybe also test...
