+++ public/js/components/Editor.js
@@ -2,7 +2,6 @@
 
 const React = require("react");
 const ReactDOM = require("react-dom");
-const Isvg = React.createFactory(require("react-inlinesvg"));
 const ImPropTypes = require("react-immutable-proptypes");
 const { bindActionCreators } = require("redux");
 const { connect } = require("react-redux");
@@ -16,53 +15,12 @@ const {
 } = require("../selectors");
 const actions = require("../actions");
 const { alignLine, onWheel } = require("../util/editor");
+const Breakpoint = React.createFactory(require("./Editor/Breakpoint"));
 
 require("codemirror/lib/codemirror.css");
 require("./Editor.css");
 require("codemirror/mode/javascript/javascript");
 require("../lib/codemirror.css");
-
-function makeMarker() {
-  let marker = document.createElement("div");
-  marker.className = "editor breakpoint";
-  ReactDOM.render(
-    React.createElement(Isvg, {
-      src: "images/breakpoint.svg#base-path___2142144446"
-    }),
-    marker
-  );
-  return marker;
-}
-
-const _Breakpoint = React.createClass({
-  propTypes: {
-    breakpoint: ImPropTypes.map,
-    editor: PropTypes.object
-  },
-
-  displayName: "Breakpoint",
-
-  componentWillMount: function() {
-    const bp = this.props.breakpoint;
-    const line = bp.getIn(["location", "line"]) - 1;
-
-    this.props.editor.setGutterMarker(line, "breakpoints", makeMarker());
-    this.props.editor.addLineClass(line, "line", "breakpoint");
-  },
-
-  componentWillUnmount: function() {
-    const bp = this.props.breakpoint;
-    const line = bp.getIn(["location", "line"]) - 1;
-
-    this.props.editor.setGutterMarker(line, "breakpoints", null);
-    this.props.editor.removeLineClass(line, "line", "breakpoint");
-  },
-
-  render: function() {
-    return null;
-  }
-});
-const Breakpoint = React.createFactory(_Breakpoint);
 
 const Editor = React.createClass({
   propTypes: {
